docs: comments to actors
