+++ collector/topic/src/main/scala/no/sysco/middleware/kafka/metadata/collector/topic/internal/TopicEventProducer.scala
@@ -15,6 +15,9 @@ object TopicEventProducer {
   def props()(implicit materializer: ActorMaterializer) = Props(new TopicEventProducer())
 }
 
+/**
+  * Publish Topic events.
+  */
 class TopicEventProducer(implicit materializer: ActorMaterializer) extends Actor {
 
   val config: Config = ConfigFactory.load() //TODO
